requirements.txt
llama-index==0.12.49
llama-index-vector-stores-chroma
llama-index-llms-ollama
llama-index-readers-file
llama-index-multi-modal-llms-ollama
chromadb~=1.0.15
tqdm~=4.67.1
python-dotenv~=1.1.1
fastapi~=0.116.1
uvicorn~=0.35.0
gradio
Pillow~=11.3.0
pillow-heif~=1.0.0
fastapi
uvicorn
llama-index
llama-index-vector-stores-postgres
llama-index-llms-ollama
llama-index-embeddings-ollama
psycopg2
sqlalchemy~=2.0.41
pgvector
psycopg2-binary
llama-index-core~=0.12.49
pydantic~=2.11.7
python-frontmatter~=1.1.0
exifread~=3.3.2